feat(error): add mongodb error management
