+++ src/app.js
@@ -3,8 +3,8 @@ import cors from 'cors';
 import dotenv from 'dotenv';
 import router from '#routes/router.js';
 import { databaseConnection } from '#utils/database.js';
-import { errorHandler } from './middlewares/error.js';
-import AppError from './utils/appError.js';
+import { errorHandler } from '#controllers/error.js';
+import AppError from '#utils/appError.js';
 
 dotenv.config();
